Remove superfluous C glue code and directly use C++ functions
